Fold bridge/formatting.py into bridge/message_drafter.py (task 1, #1035)
- Move linkify_references and linkify_references_from_session into message_drafter
- Update tools/send_telegram.py, tools/valor_telegram.py, test imports
- Rename tests/unit/test_formatting.py -> test_message_drafter_linkify.py
- Fix test mock path: bridge.routing.load_config (where it's imported from)
- Delete bridge/formatting.py (76 lines; zero net new lines)
- Keep _linkify_references as a private backward-compat alias so internal
  callers (response.py, _compose_structured_draft) don't need to change
