+++ bridge/message_drafter.py
@@ -982,21 +982,75 @@ def _parse_classification_response(raw: str) -> ClassificationResult | None:
     return result
 
 
-def _linkify_references(text: str, session) -> str:
+def linkify_references(text: str, project_key: str | None = None) -> str:
     """Convert plain PR #N and Issue #N references to markdown links.
 
-    Delegates to bridge.formatting.linkify_references_from_session for the
-    actual implementation. This wrapper preserves the existing call signature.
+    Uses the project_key to look up the GitHub org/repo from the registered
+    project config. If no project config is found or the text already
+    contains markdown links for a reference, it is left unchanged.
+
+    Originally in bridge/formatting.py; folded into bridge/message_drafter.py
+    per the message-drafter consolidation (plan #1035 Part A).
 
     Args:
-        text: The draft text potentially containing PR #N or Issue #N
-        session: AgentSession with project_key field
+        text: The text potentially containing PR #N or Issue #N references.
+        project_key: Project key for GitHub org/repo lookup. If None, text
+            is returned unchanged.
 
     Returns:
         Text with plain references converted to markdown links.
     """
-    from bridge.formatting import linkify_references_from_session
+    from bridge.routing import load_config
 
+    if not text or not project_key or not str(project_key).strip():
+        return text
+
+    try:
+        all_projects = load_config().get("projects", {})
+        config = all_projects.get(str(project_key), {})
+        github_config = config.get("github", {})
+        org = github_config.get("org")
+        repo = github_config.get("repo")
+    except Exception:
+        return text
+
+    if not org or not repo:
+        return text
+
+    base_url = f"https://github.com/{org}/{repo}"
+
+    # Negative lookbehind for [ ensures we don't double-link already-linked refs.
+    text = re.sub(
+        r"(?<!\[)PR #(\d+)(?!\])",
+        lambda m: f"[PR #{m.group(1)}]({base_url}/pull/{m.group(1)})",
+        text,
+    )
+    text = re.sub(
+        r"(?<!\[)Issue #(\d+)(?!\])",
+        lambda m: f"[Issue #{m.group(1)}]({base_url}/issues/{m.group(1)})",
+        text,
+    )
+    return text
+
+
+def linkify_references_from_session(text: str, session) -> str:
+    """Convenience wrapper that extracts project_key from a session object.
+
+    Args:
+        text: The text to linkify.
+        session: Object with a project_key attribute (e.g., AgentSession).
+
+    Returns:
+        Text with plain references converted to markdown links.
+    """
+    if not session:
+        return text
+    project_key = getattr(session, "project_key", None)
+    return linkify_references(text, project_key)
+
+
+def _linkify_references(text: str, session) -> str:
+    """Backward-compat alias for linkify_references_from_session."""
     return linkify_references_from_session(text, session)
